Add initial on-call rotation policy
This page is a formalization of what @seth-meltano proposed, and what the Engineering Team discussed, as best I can recall. Some additions (e.g. how new hires are added, the ability to trade shifts, etc.) and clarifications were not previously discussed.

Feedback/criticism is strongly encouraged - this policy is by us, and for us. No one enjoys being on-call, but it's a necessity for Meltano, so we should make it the best it can be.